requirements.txt
google-genai
typer
rich
pandas
python-dotenv
scikit-learn
